Moved job_queue demo to it's own file
